sw_apps:zynqmp_fsbl:Added #ifdef XFSBL_BS at missing
Provided minor enhancements for bitstream secure
code, by adding C header to newly added .h file,
added #ifdef at missing places.

Acked-for-series: Krishna Chaitanya Patakamuri <kpataka@xilinx.com>

+++ lib/sw_apps/zynqmp_fsbl/src/xfsbl_misc.c
@@ -755,7 +755,7 @@ END:
 *		simple DMA.
 *
 ******************************************************************************/
-u32 XFsbl_AdmaCopy(void * DestPtr, void * SrcPtr, u64 Size)
+u32 XFsbl_AdmaCopy(void * DestPtr, void * SrcPtr, u32 Size)
 {
 	u32 RegVal;
 	u64 SrcAddr = (UINTPTR)SrcPtr;
